have BasePauselessRealtimeIngestionTest inherit BaseClusterIntegrationTestSet
